Implement a way to show details of a specific version
- Don't show *all* the results, as we don't know yet if that can
  be used to overfit a model to just meet the requirements
- Redirect to a version's detail page after upload
- User can click the refresh button to check for updates
- Link to version detail page from submissions page

Ref https://github.com/2i2c-org/frx-challenges/issues/189
