+++ frx_challenges/web/templates/submission/detail.html
@@ -64,7 +64,7 @@
                     <tbody>
                         {% for v in versions %}
                             <tr>
-                                <td>{{ v.id }}</td>
+                                <td><a href="{% url 'versions-view' v.id %}">{{ v.id }}</a></td>
                                 <td>{{ v.filename }}</td>
                                 <td>{{ v.date_created|date:"c" }}</td>
                                 <td>{{ v.latest_evaluation.status }}</td>
